feat: actualiza estructura de react
