--- client/.env
@@ -1,1 +1,0 @@
-VITE_API_BASE_URL=http://localhost:8000
