+++ notes-frontend-app/eslint.config.js
