+++ notes-frontend-app/src/containers/errors/error_404.tsx
@@ -1,0 +1,8 @@
+function Error404(){
+    return (
+        <div>
+        <h1>Error404</h1>
+        </div>
+    )
+}
+export default Error404
