+++ notes-frontend-app/src/containers/pages/Home.tsx
@@ -1,0 +1,8 @@
+function Home(){
+    return (
+        <div>
+        <h1>Home</h1>
+        </div>
+    )
+}
+export default Home
