+++ notes-frontend-app/src/redux/reducers/index.js
@@ -1,0 +1,5 @@
+import { combineReducers } from 'redux';
+
+export default combineReducers({
+
+})
