+++ notes-frontend-app/src/styles/index.css
