+++ notes-frontend-app/src/vite-env.d.ts
