+++ notes-frontend-app/vite.config.ts
